requirements.txt
aiofiles==24.1.0
aiogram==3.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
audioop-lts==0.2.1
beautifulsoup4==4.13.4
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
dashscope==1.24.0
deep-translator==1.11.4
dotenv==0.9.9
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
langdetect==1.0.9
magic-filter==1.0.12
multidict==6.6.3
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
PyOgg==0.6.14a1
python-dotenv==1.1.1
python-telegram-bot==22.3
pytz==2025.2
requests==2.32.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SpeechRecognition==3.14.3
standard-aifc==3.13.0
standard-chunk==3.13.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzlocal==5.3.1
urllib3==2.5.0
websocket-client==1.8.0
yarl==1.20.1
